requirements.txt
streamlit>=1.53.0
python-dotenv>=1.0.0
pandas>=2.3.0
numpy>=2.2.0
openpyxl>=3.1.0
requests>=2.32.0
google-genai>=1.60.0
google-cloud-storage>=2.10.0
google-auth>=2.0.0
google-api-python-client>=2.0.0
toml>=0.10.0
boto3>=1.28.0
plotly>=6.5.0
rapidfuzz>=3.14.0
reportlab>=4.0.0
statsmodels>=0.14.0
scipy>=1.11.0
pypdf>=3.0.0